+++ modules/contracts/src.ts/tests/utils.ts
@@ -1,5 +1,6 @@
+import { BigNumber } from "@ethersproject/bignumber";
 import { AddressZero } from "@ethersproject/constants";
-import { BigNumber, Contract } from "ethers";
+import { Contract } from "@ethersproject/contracts";
 
 import { createChannel, deployContracts } from "../actions";
 import { AddressBook, getAddressBook } from "../addressBook";
@@ -26,15 +27,11 @@ export const getUnsetupChannel = async (_addressBook?: AddressBook): Promise<Con
     ["TestChannel", []],
     ["TestChannelFactory", ["TestChannel"]],
   ]);
-  const factory = addressBook.getContract("TestChannelFactory");
-  const doneBeingCreated: Promise<string> = new Promise(res => {
-    // NOTE: this takes kind of a long time to resolve.. is there any way to speed it up?
-    factory.once(factory.filters.ChannelCreation(), res);
-  });
+  const testFactory = addressBook.getContract("TestChannelFactory");
   const chainId = (await alice.provider.getNetwork()).chainId.toString();
-  const tx = await factory.createChannelWithoutSetup(alice.address, bob.address, chainId);
+  const channelAddress = await testFactory.getChannelAddress(alice.address, bob.address, chainId);
+  const tx = await testFactory.createChannelWithoutSetup(alice.address, bob.address, chainId);
   await tx.wait();
-  const channelAddress = await doneBeingCreated;
   // Save this channel address in case we need it later
   addressBook.setEntry(`VectorChannel-${alice.address.substring(2, 6)}-${bob.address.substring(2, 6)}`, {
     address: channelAddress,